Add files via upload
This code checks if a given block configuration hits all the target coordinates.

This code handles multiple lazers and Reflect and Opaque block types.
Inputs:
1) Lazer coordinates and directions for each lazer. Append to 'lazerList'
2) Block Details: Big coordinate system's coordinates and type. Append to 'dictionary
3) Target coordinates.

The main part of the code will select a lazer from the list of lazers. And it will traverse. If it hits a Reflect block, it will change the direction accordingly. If it hits an opaque block, it will stop at that point.

This will be done for each lazer in the list.

The result will be a list of all the coordinates covered by the lazer. If the targets are a subset of that list, it is a valid configuration.

NEXT:
1) Refract. (There is a small problem to this. Will discuss it later)
2) Make it object oriented. Classes.
3) Integrate with the parsed file and GUI.